+++ packages/logic-core/src/data/solver/backtrack/symbols/dart.ts
@@ -14,6 +14,7 @@ export default class DartBTModule extends BTModule {
   public instr: DartSymbol;
 
   private cachedCheckResult?: CheckResult;
+  private cachedState?: { completed: number; empty: number };
 
   public constructor(instr: DartSymbol) {
     super();
@@ -49,8 +50,14 @@ export default class DartBTModule extends BTModule {
 
     if (completed + empty < this.instr.number) return false;
 
-    if (!this.cachedCheckResult)
+    if (
+      !this.cachedCheckResult ||
+      this.cachedState?.completed !== completed ||
+      this.cachedState?.empty !== empty
+    ) {
       this.cachedCheckResult = this.buildCheckAndRating(grid);
+      this.cachedState = { completed, empty };
+    }
 
     return this.cachedCheckResult;
   }